+++ lib/src/components/notification/item.vue
@@ -1,0 +1,16 @@
+<template lang="pug">
+  .common-notification-component
+    div notification
+</template>
+
+<script>
+export default {
+  name: 'common-notification-component'
+}
+</script>
+
+<style lang="scss" scoped>
+  .common-notification-component {
+    //
+  }
+</style>